requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
joblib==1.3.2
pandas==2.2.1
scikit-learn==1.4.1.post1
reportlab==4.1.0
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
